+++ app/src/main/kotlin/model/graph/classes/DirectedWeightedGraph.kt
@@ -4,12 +4,12 @@ import model.graph.abstractClasses.AbstractGraph
 import model.graph.dataClasses.Vertex
 import model.graph.dataClasses.WeightedEdge
 
-class DirectedWeightedGraph<V : Comparable<V>> : AbstractGraph<Int, V>() {
+class DirectedWeightedGraph<V : Comparable<V>> : AbstractGraph<Long, V>() {
     fun addEdge(
         from: Vertex<V>,
         to: Vertex<V>,
-        weight: Int,
-    ): Int? = addEdge(WeightedEdge(nextEdgeId++, from, to, weight))
+        weight: Long,
+    ): Long? = addEdge(WeightedEdge(nextEdgeId++, from, to, weight))
 
     override fun getNeighbors(vertex: Vertex<V>): List<Vertex<V>> {
         if (vertex !in _vertices.values) return emptyList()
