+++ app/src/main/kotlin/model/graph/classes/UndirectedUnweightedGraph.kt
@@ -4,9 +4,9 @@ import model.graph.abstractClasses.AbstractGraph
 import model.graph.dataClasses.UnweightedEdge
 import model.graph.dataClasses.Vertex
 
-class UndirectedUnweightedGraph<V : Comparable<V>> : AbstractGraph<Int, V>() {
+class UndirectedUnweightedGraph<V : Comparable<V>> : AbstractGraph<Long, V>() {
     fun addEdge(
         from: Vertex<V>,
         to: Vertex<V>,
-    ): Int? = addEdge(UnweightedEdge(nextEdgeId++, from, to))
+    ): Long? = addEdge(UnweightedEdge(nextEdgeId++, from, to))
 }
