+++ app/src/main/kotlin/model/graph/classes/UndirectedWeightedGraph.kt
@@ -4,10 +4,10 @@ import model.graph.abstractClasses.AbstractGraph
 import model.graph.dataClasses.Vertex
 import model.graph.dataClasses.WeightedEdge
 
-class UndirectedWeightedGraph<V : Comparable<V>> : AbstractGraph<Int, V>() {
+class UndirectedWeightedGraph<V : Comparable<V>> : AbstractGraph<Long, V>() {
     fun addEdge(
         from: Vertex<V>,
         to: Vertex<V>,
-        weight: Int,
-    ): Int? = addEdge(WeightedEdge(nextEdgeId++, from, to, weight))
+        weight: Long,
+    ): Long? = addEdge(WeightedEdge(nextEdgeId++, from, to, weight))
 }
