+++ app/src/main/kotlin/model/graph/dataClasses/WeightedEdge.kt
@@ -6,5 +6,5 @@ data class WeightedEdge<E : Comparable<E>, V : Comparable<V>>(
     override val id: E,
     override val from: Vertex<V>,
     override val to: Vertex<V>,
-    val weight: Int,
+    val weight: Long,
 ) : Edge<E, V>
